refactor redfin parser and move logic to a different file
